+++ CHANGELOG.md
@@ -5,7 +5,6 @@
 - Add on more known 2D plotting functions
 - Flesh out testing of 3D plotting
 - AxesAnaglyph.imshow_stereo() nonlinear color mapping methods
-- 3D plot anaglyphs
 
 ### Known issues:
 - Inconsistent coloring across redraws if not specified
@@ -14,6 +13,8 @@
 
 ## [Unreleased]
 ### Added    
+* `AxesAnaglyph3D` class for stereoscopic 3D red-cyan anaglyph plots
+* `StereoSquare3D` now fills its anaglyph cell using `AxesAnaglyph3D`
 ### Changed    
 ### Removed